Update library project to use “implementation” instead of “compile”
For quite some time now (gradle 3.0), gradle has added support for “api”
and “implementation” keywords instead of “compile” which enable one to
choose whether the library should propagate dependencies to its
dependents (api) or merely use it for its own purposes and not make them
available for modules depending on it (implementation).

For the library, this might result in breaking changes for some
dependents of it as “implementation” no longer automatically adds the
support libraries used by the this library to their build path, making
them readily available for use in their app as well without defining
them as dependencies on their own.